+++ templates/index.html
@@ -23,7 +23,7 @@
     {% for mytask in mytasks %}
       <p class="task_text">&#10143; {{ mytask.tdtask }}</p>
       <p class="task_meta">Due date: <strong>{{ mytask.lastday }}</strong> |
-      {% if mytask.complete == false %} <input type="checkbox" name="iscomplete" onclick='window.location.assign("{{ url_for("done", id=mytask.key) }}")'/><em>Mark as complete</em>{% else %}<em>Task completed</em>{% endif %}</p>
+      {% if mytask.complete == false %} <input type="checkbox" name="iscomplete" onclick='window.location.assign("{{ url_for("done", t_id=mytask.key) }}")'/><em>Mark as complete</em>{% else %}<em>Task completed</em>{% endif %}</p>
     {% endfor %}
   </article>
 </section>
